+++ src/cloudant/errors.py
@@ -49,3 +49,30 @@ class CloudantArgumentError(CloudantException):
     """
     def __init__(self, msg, code=400):
         super(CloudantArgumentError, self).__init__(msg, code)
+
+class ResultException(CloudantException):
+    """
+    Provides a way to issue Cloudant Python client library result specific
+    exceptions.
+
+    :param int code: A code value used to identify the result exception.
+        Defaults to 100.
+    :param args: A list of arguments used to format the exception message.
+    """
+
+    def __init__(self, code=100, *args):
+        messages = {
+            100:'A general result exception was raised.',
+            101:'Failed to interpret the argument {0} as a valid key value or as a valid slice.',
+            102:'Cannot use {0} when performing key access or key slicing.  Found {1}.',
+            103:'Cannot use {0} for iteration.  Found {1}.',
+            104:'Invalid page_size: {0}'
+        }
+
+        try:
+            msg = messages[code].format(*args)
+        # pylint: disable=broad-except
+        except Exception:
+            code = 100
+            msg = messages[code]
+        super(ResultException, self).__init__(msg, code)