+++ src/cloudant/result.py
@@ -19,7 +19,7 @@ import json
 from collections import Sequence
 
 from ._2to3 import STRTYPE, UNITYPE, NONETYPE, iteritems_
-from .errors import CloudantArgumentError
+from .errors import CloudantArgumentError, ResultException
 
 ARG_TYPES = {
     'descending': (bool,),
@@ -129,40 +129,80 @@ def type_or_none(typerefs, value):
     """
     return isinstance(value, typerefs) or value is None
 
-class Result(object):
+class ResultByKey(object):
     """
-    Provides a sliceable and iterable interface to result collections.
-    A Result object is instantiated with a raw data callable reference
-    such as the database API convenience method
-    :func:`~cloudant.database.CouchDatabase.all_docs` or the View
-    :func:`~cloudant.views.View.__call__` callable, both used to retrieve data.
-    A Result object can also use optional extra arguments for result
-    customization and supports efficient, paged iteration over the result
-    collection to avoid large result data from adversely affecting memory.
-
-    In Python, slicing returns by value, whereas iteration will yield
-    elements of the sequence.  This means that slicing will perform better
-    for smaller data collections, whereas iteration will be more
-    efficient for larger data collections.
+    Provides a wrapper for a value used to retrieve records from a result
+    collection based on an actual document key value.  This comes in handy when
+    the document key value is an ``int``.
 
     For example:
 
     .. code-block:: python
 
-        # Access by key:
-        result['key'] # get all records matching key
+        result = Result(callable)
+        result[ResultByKey(9)]   # gets records where the key matches 9
+        # as opposed to:
+        result[9]                # gets the 10th record of the result collection
 
-        # Slicing by startkey/endkey:
-        result[['2013','10']:['2013','11']] # results between compound keys
-        result['2013':'2014']               # results between string keys
-        result['2013':]                     # all results after key
-        result[:'2014']                     # all results up to key
+        :param value: A value representing a Result key.
+    """
+    def __init__(self, value):
+        self._value = value
 
-        # Slicing by value:
-        result[100:200] # results between the 100th and the 200th result
-        result[:200]    # results up to the 200th result
-        result[100:]    # results after 100th result
-        result[:]       # all results
+    def __call__(self):
+        return self._value
+
+class Result(object):
+    """
+    Provides a key accessible, sliceable, and iterable interface to result
+    collections.  A Result object is constructed with a raw data callable
+    reference such as the database API convenience method
+    :func:`~cloudant.database.CouchDatabase.all_docs` or the View
+    :func:`~cloudant.views.View.__call__` callable, used to retrieve data.
+    A Result object can also use optional extra arguments for result
+    customization and supports efficient, paged iteration over the result
+    collection to avoid large result data from adversely affecting memory.
+
+    In Python, slicing returns by value, whereas iteration will yield
+    elements of the sequence.  This means that individual key access and slicing
+    will perform better for smaller data collections, whereas iteration will
+    be more efficient for larger data collections.
+
+    For example:
+
+    .. code-block:: python
+
+        # Key access:
+
+        # Access by index value:
+        result = Result(callable)
+        result(9)                 # skip first 9 records and get 10th
+
+        # Access by key value:
+        result = Result(callable)
+        result['foo']             # get records matching 'foo'
+        result[ResultByKey(9)]    # get records matching 9
+
+        # Slice access:
+
+        # Access by index slices:
+        result = Result(callable)
+        result[100: 200]          # get records after the 100th and up to and including the 200th
+        result[: 200]             # get records up to and including the 200th
+        result[100: ]             # get all records after the 100th
+        result[: ]                # get all records
+
+        # Access by key slices:
+        result = Result(callable)
+        result['bar':'foo']       # get records between and including 'bar' and 'foo'
+        result['foo':]            # get records after and including 'foo'
+        result[:'foo']            # get records up to and including 'foo'
+
+        result[['foo', 10]:
+               ['foo', 11]]       # Complex key access and slicing works the same as simple keys
+
+        result[ResultByKey(5):
+               ResultByKey(10)]   # key slice access of integer keys
 
         # Iteration:
 
@@ -176,99 +216,193 @@ class Result(object):
         for i in result:
             print i
 
-        # Iterate over the entire result collection,
-        # including documents and in batches of a 1000.
+        # Iterate over the entire result collection in batches of 1000, including documents.
         result = Result(callable, include_docs=True, page_size=1000)
         for i in result:
             print i
 
-        :param method_ref: A reference to the method or callable that returns
-            the JSON content result to be wrapped.
-        :param options: See :func:`~cloudant.views.View.make_result` for a
-            list of valid result customization options.
+    Note: Since Result object key access, slicing, and iteration use query
+    parameters behind the scenes to handle their processing, some query
+    parameters are not permitted as part of a Result customization,
+    depending on whether key access, slicing, or iteration is being performed.
+
+    Such as:
+
+    +-------------------------------+-----------------------------------------------------------+
+    | Access/Slicing by index value | No restrictions                                           |
+    +-------------------------------+-----------------------------------------------------------+
+    | Access/Slicing by key value   | ``key``, ``keys``, ``startkey``, ``endkey`` not permitted |
+    +-------------------------------+-----------------------------------------------------------+
+    | Iteration                     | ``limit``, ``skip`` not permitted                         |
+    +-------------------------------+-----------------------------------------------------------+
+
+    :param method_ref: A reference to the method or callable that returns
+        the JSON content result to be wrapped as a Result.
+    :param options: See :func:`~cloudant.views.View.make_result` for a
+        list of valid result customization options.
     """
     def __init__(self, method_ref, **options):
         self.options = options
         self._ref = method_ref
         self._page_size = options.pop('page_size', 100)
 
-    def __getitem__(self, key):
+    def __getitem__(self, arg):
         """
         Provides Result key access and slicing support.
 
-        See :class:`~cloudant.result.Result` for key access and slicing
-        examples.
+        An ``int`` argument will be interpreted as a ``skip`` and then a get of
+        the next record.  For example ``[100]`` means skip the first 100 records
+        and then get the next record.
 
-        :param key:  Can be either a single value as a ``str`` or ``list``
-            which will be passed as the key to the query for entries matching
-            that key or slice.  Slices with integers will be interpreted as
-            ``skip:limit-skip`` style pairs.  For example ``[100:200]`` means
-            skip the first 100 records then get up to and including the 200th
-            record so that you get the range between the supplied slice values.
-            Slices with strings/lists will be interpreted as startkey/endkey
-            style keys.
+        A ``str``, ``list`` or :class:`~cloudant.result.ResultByKey` argument
+        will be interpreted as a ``key`` and then get all records that match the
+        given key.  For example ``['foo']`` will get all records that match
+        the key 'foo'.
 
-        :returns: Rows data in JSON format
+        An ``int`` slice argument will be interpreted as a ``skip:limit-skip``
+        style pair.  For example ``[100: 200]`` means skip the first 100 records
+        then get up to and including the 200th record so that you get the range
+        between the supplied slice values.
+
+        A slice argument that contains ``str``, ``list``, or
+        :class:`~cloudant.result.ResultByKey` will be interpreted as a
+        ``startkey: endkey`` style pair.  For example ``['bar': 'foo']`` means
+        get the range of records where the keys are between and including
+        'bar' and 'foo'.
+
+        See :class:`~cloudant.result.Result` for more detailed key access and
+        slicing examples.
+
+        :param arg:  A single value representing a key or a pair of values
+            representing a slice.  The argument value(s) can be ``int``,
+            ``str``, ``list`` (in the case of complex keys), or
+            :class:`~cloudant.result.ResultByKey`.
+
+        :returns: Rows data as a list in JSON format
         """
-        if isinstance(key, STRTYPE):
-            data = self._ref(key=key, **self.options)
-            return self._parse_data(data)
+        data = None
+        if isinstance(arg, int):
+            data = self._handle_result_by_index(arg)
+        elif isinstance(arg, STRTYPE) or isinstance(arg, list):
+            data = self._handle_result_by_key(arg)
+        elif isinstance(arg, ResultByKey):
+            data = self._handle_result_by_key(arg())
+        elif isinstance(arg, slice):
+            # slice is entire result set - no additional processing required
+            if arg.start is None and arg.stop is None:
+                data = self._ref(**self.options)
+            # key slice identified
+            elif (type_or_none((STRTYPE, list, ResultByKey), arg.start) and
+                  type_or_none((STRTYPE, list, ResultByKey), arg.stop)):
+                data = self._handle_result_by_key_slice(arg)
+            # index slice identified
+            elif (type_or_none(int, arg.start) and
+                  type_or_none(int, arg.stop)):
+                data = self._handle_result_by_idx_slice(arg)
+        if data is None:
+            raise ResultException(101, arg)
+        return self._parse_data(data)
 
-        if isinstance(key, list):
-            data = self._ref(key=key, **self.options)
-            return self._parse_data(data)
+    def _handle_result_by_index(self, idx):
+        """
+        Handle processing when the result argument provided is an integer.
+        """
+        if idx < 0:
+            return None
+        opts = dict(self.options)
+        skip = opts.pop('skip', 0)
+        limit = opts.pop('limit', None)
+        _validate('skip', skip)
+        _validate('limit', limit)
+        if limit is not None and idx >= limit:
+            # Result is out of range
+            return dict()
+        return self._ref(skip=skip+idx, limit=1, **opts)
 
-        if isinstance(key, slice):
-            # slice is startkey and endkey if str or array
-            str_or_none_start = type_or_none((STRTYPE, list), key.start)
-            str_or_none_stop = type_or_none((STRTYPE, list), key.stop)
-            if str_or_none_start and str_or_none_stop:
-                # startkey/endkey
-                if key.start is not None and key.stop is not None:
-                    data = self._ref(
-                        startkey=key.start,
-                        endkey=key.stop,
-                        **self.options
-                    )
-                if key.start is not None and key.stop is None:
-                    data = self._ref(startkey=key.start, **self.options)
-                if key.start is None and key.stop is not None:
-                    data = self._ref(endkey=key.stop, **self.options)
-                if key.start is None and key.stop is None:
-                    data = self._ref(**self.options)
-                return self._parse_data(data)
-            # slice is skip:skip+limit if ints
-            int_or_none_start = type_or_none(int, key.start)
-            int_or_none_stop = type_or_none(int, key.stop)
-            if int_or_none_start and int_or_none_stop:
-                if key.start is not None and key.stop is not None:
-                    limit = key.stop - key.start
-                    data = self._ref(
-                        skip=key.start,
-                        limit=limit,
-                        **self.options
-                    )
-                if key.start is not None and key.stop is None:
-                    data = self._ref(skip=key.start, **self.options)
-                if key.start is None and key.stop is not None:
-                    data = self._ref(limit=key.stop, **self.options)
-                # both None case handled already
-                return self._parse_data(data)
-        msg = (
-            'Failed to interpret the argument {0} '
-            'as a valid key value or as a valid slice.'
-        ).format(key)
-        raise CloudantArgumentError(msg)
+    def _handle_result_by_key(self, key):
+        """
+        Handle processing when the result argument provided is a document key.
+        """
+        invalid_options = ('key', 'keys', 'startkey', 'endkey')
+        if any(x in invalid_options for x in self.options):
+            raise ResultException(102, invalid_options, self.options)
+        return self._ref(key=key, **self.options)
+
+    def _handle_result_by_idx_slice(self, idx_slice):
+        """
+        Handle processing when the result argument provided is an index slice.
+        """
+        opts = dict(self.options)
+        skip = opts.pop('skip', 0)
+        limit = opts.pop('limit', None)
+        _validate('skip', skip)
+        _validate('limit', limit)
+        start = idx_slice.start
+        stop = idx_slice.stop
+        data = None
+        if (start is not None and stop is not None and
+                start >= 0 and stop >= 0 and start <= stop):
+            if limit is not None:
+                if start >= limit:
+                    # Result is out of range
+                    return dict()
+                elif stop > limit:
+                    # Ensure that slice does not extend past original limit
+                    return self._ref(skip=skip+start, limit=limit-start, **opts)
+            data = self._ref(skip=skip+start, limit=stop-start, **opts)
+        elif start is not None and stop is None and start >= 0:
+            if limit is not None:
+                if start >= limit:
+                    # Result is out of range
+                    return dict()
+                # Ensure that slice does not extend past original limit
+                data = self._ref(skip=skip+start, limit=limit-start, **opts)
+            else:
+                data = self._ref(skip=skip+start, **opts)
+        elif start is None and stop is not None and stop >= 0:
+            if limit is not None and stop > limit:
+                # Ensure that slice does not extend past original limit
+                data = self._ref(skip=skip, limit=limit, **opts)
+            else:
+                data = self._ref(skip=skip, limit=stop, **opts)
+        return data
+
+    def _handle_result_by_key_slice(self, key_slice):
+        """
+        Handle processing when the result argument provided is a key slice.
+        """
+        invalid_options = ('key', 'keys', 'startkey', 'endkey')
+        if any(x in invalid_options for x in self.options):
+            raise ResultException(102, invalid_options, self.options)
+
+        if isinstance(key_slice.start, ResultByKey):
+            start = key_slice.start()
+        else:
+            start = key_slice.start
+
+        if isinstance(key_slice.stop, ResultByKey):
+            stop = key_slice.stop()
+        else:
+            stop = key_slice.stop
+
+        if (start is not None and stop is not None and
+                isinstance(start, type(stop))):
+            data = self._ref(startkey=start, endkey=stop, **self.options)
+        elif start is not None and stop is None:
+            data = self._ref(startkey=start, **self.options)
+        elif start is None and stop is not None:
+            data = self._ref(endkey=stop, **self.options)
+        else:
+            data = None
+        return data
 
     def __iter__(self):
         """
         Provides iteration support, primarily for large data collections.
-        The iterator uses the skip/limit parameters to consume data in chunks
-        controlled by the ``page_size`` setting and retrieves a batch of data
-        from the result collection and then yields each element.  Since the
-        iterator uses the skip/limit parameters to perform the iteration,
-        ``skip`` and ``limit`` cannot be included as part of the original result
-        customization options.
+        The iterator uses the ``skip`` and ``limit`` options to consume
+        data in chunks controlled by the ``page_size`` option.  It retrieves
+        a batch of data from the result collection and then yields each
+        element.
 
         See :func:`~cloudant.views.View.make_result` for a list of valid
         result customization options.
@@ -277,25 +411,25 @@ class Result(object):
 
         :returns: Iterable data sequence
         """
-        if 'skip' in self.options:
-            msg = 'Cannot use skip for iteration'
-            raise CloudantArgumentError(msg)
-        if 'limit' in self.options:
-            msg = 'Cannot use limit for iteration'
-            raise CloudantArgumentError(msg)
-        if self._page_size <= 0:
-            msg = 'Invalid page_size: {0}'.format(self._page_size)
-            raise CloudantArgumentError(msg)
+        invalid_options = ('skip', 'limit')
+        if any(x in invalid_options for x in self.options):
+            raise ResultException(103, invalid_options, self.options)
+
+        try:
+            if int(self._page_size) <= 0:
+                raise ResultException(104, self._page_size)
+        except ValueError:
+            raise ResultException(104, self._page_size)
 
         skip = 0
         while True:
             response = self._ref(
-                limit=self._page_size,
+                limit=int(self._page_size),
                 skip=skip,
                 **self.options
             )
             result = self._parse_data(response)
-            skip += self._page_size
+            skip += int(self._page_size)
             if len(result) > 0:
                 for row in result:
                     yield row